requirements.txt
requests
selenium
regex
SQLAlchemy
alembic
progressbar2
